docs: add column limit specification to coding style
Google's Java coding style sets column limit at 100. But because of the
increased block indentation, increased indent continuation lines limit,
and descriptive method/variable names, the column limit is reached
easily. For this project, increase the column limit to 120.

This will increase readability of the code and decrease the overall line
count.

Signed-off-by: 김명종 <mj610.kim@gmail.com>