refactor: extract Cmd-specific methods to Cmd enum class
Previously, cmd parsing was done manually with a switch statement. This
can cause bugs later when even a single character is changed. To make it
clearer which values are allowed for the cmd, use an enum class. Each
value of the enum class has different "run()" behavior, so define it as
an abstract method in the enum class itself.

Signed-off-by: 김명종 <mj610.kim@gmail.com>