ci: add github actions TC for name substring cases
There were cases in which a person that should not have been included
was included in DEL, SCH, and MOD commands. This was because current
implementation of DB returns all employees that contain the query name
as a substring. This has to be fixed, and this TC will verify the fix.

Signed-off-by: 김명종 <mj610.kim@gmail.com>